robomonitor/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
boto3==1.28.24
botocore==1.31.24
celery==5.3.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-redis==5.3.0
django-tables2==2.6.0
django-timezone-field==5.1
django4-background-tasks==1.2.7
dnspython @ git+https://github.com/rthalley/dnspython.git@a22644d7ced90ee796592622eebd9629377b39ba
frozenlist==1.4.0
idna==3.4
install==1.3.5
jmespath==1.0.1
kombu==5.3.1
multidict==6.0.4
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
redis==4.6.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
